+++ src/main/java/org/truffleruby/core/kernel/TruffleKernelNodes.java
@@ -53,13 +53,17 @@ import com.oracle.truffle.api.dsl.ImportStatic;
 import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.FrameDescriptor;
+import com.oracle.truffle.api.frame.FrameInstance.FrameAccess;
 import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.frame.FrameSlotKind;
 import com.oracle.truffle.api.frame.FrameUtil;
 import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.IndirectCallNode;
+import com.oracle.truffle.api.nodes.Node;
+import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.profiles.ConditionProfile;
+import com.oracle.truffle.api.Truffle;
 
 @CoreModule("Truffle::KernelOperations")
 public abstract class TruffleKernelNodes {
@@ -320,6 +324,38 @@ public abstract class TruffleKernelNodes {
         @Specialization
         protected Object storage(VirtualFrame frame) {
             return callerVariablesNode.execute(frame);
+        }
+    }
+
+    /* When getting special variables from the wrong side of a C call we know it's going to be slow. */
+    @Primitive(name = "slow_caller_special_variables")
+    public abstract static class GetSlowCallerSpecialVariableStorage extends PrimitiveArrayArgumentsNode {
+
+        @Child GetSpecialVariableStorage getStorageNode = GetSpecialVariableStorage.create();
+
+        @Specialization
+        @TruffleBoundary
+        protected Object storage() {
+            return getStorageNode.execute(Truffle.getRuntime().iterateFrames(frameInstance -> {
+                final Node callNode = frameInstance.getCallNode();
+
+                if (callNode != null) {
+                    final RootNode rootNode = callNode.getRootNode();
+                    // Skip Ruby frames in cext.rb file since they are implementing methods which are implemented
+                    // with C in MRI, and therefore are also implicitly skipped when when looking up the block passed
+                    // to a C API function.
+                    if (rootNode instanceof RubyRootNode &&
+                            rootNode.getSourceSection().isAvailable() &&
+                            !rootNode.getSourceSection().getSource().getName().endsWith("cext.rb") &&
+                            !rootNode.getSourceSection().getSource().getName().endsWith("cext_ruby.rb") &&
+                            !getContext().getCoreLibrary().isSend(
+                                    RubyArguments.tryGetMethod(frameInstance.getFrame(FrameAccess.READ_ONLY)))) {
+                        return frameInstance.getFrame(FrameAccess.MATERIALIZE).materialize();
+                    }
+                }
+
+                return null;
+            }));
         }
     }